interview_agents_api/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-multipart

langchain-core
langchain-community
langchain-openai
langchain_groq
langchain-huggingface
langgraph
crewai
crewai-tools
sentence_transformers
torch
transformers
sentencepiece
accelerate
pypdf
python-dotenv
pymongo
requests
faiss-cpu

httpx==0.28.1